refact(store): mockCmsService --> cmsService
