+++ apps/store/src/components/ProductPage/ProductPage.types.ts
@@ -1,4 +1,4 @@
-import { CmsProduct } from '@/services/mockCmsService'
+import { CmsProduct } from '@/services/cms'
 import { Product } from '@/services/mockProductService'
 
 export type ProductPageProps = {
