+++ apps/store/src/pages/products/[product].tsx
@@ -3,7 +3,7 @@ import { LayoutWithMenu } from '@/components/LayoutWithMenu/LayoutWithMenu'
 import { ProductPage } from '@/components/ProductPage/ProductPage'
 import { ProductPageProps } from '@/components/ProductPage/ProductPage.types'
 import { getLocale } from '@/lib/l10n/getLocale'
-import { getProductByMarketAndSlug } from '@/services/mockCmsService'
+import { Cms } from '@/services/cms'
 import { getProductByMarketAndName } from '@/services/mockProductService'
 
 const NextProductPage: NextPageWithLayout<ProductPageProps> = (props: ProductPageProps) => {
@@ -22,7 +22,7 @@ export const getServerSideProps: GetServerSideProps<ProductPageProps> = async (c
     }
   }
 
-  const cmsProduct = getProductByMarketAndSlug(localeData.marketLabel, slug)
+  const cmsProduct = await Cms.getProductByMarketAndSlug(localeData.marketLabel, slug)
 
   if (!cmsProduct) {
     return {
